requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.23.0
numpy==2.0.1
python-dateutil==2.9.0.post0
